requirements.txt
asgiref==3.5.2
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
coverage
Django==4.1
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-crispy-forms==1.13.0
django-hCaptcha==0.1.0
django-otp==1.1.3
django-photologue==3.17
django-sortedm2m==3.1.1
django-summernote==0.8.20.0
django-taggit==6.1.0
ExifRead==3.0.0
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
mysqlclient==2.1.0
packaging==21.3
pillow==10.4.0
pyparsing==3.0.6
python-decouple==3.5
python-dotenv==0.19.2
python-slugify==8.0.4
qrcode==7.3.1
requests==2.27.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
urllib3==1.26.7
webencodings==0.5.1
wheel==0.44.0
